对State模式Demo中的各个状态类进行重构，修整GUI部分逻辑代码

+++ State/VendingMachinesDemo/src/edu/zjut/model/SaleMachine.java
@@ -2,7 +2,7 @@ package edu.zjut.model;
 
 public class SaleMachine {
 	
-	public String buy(String type, int money) {
+	public String buy(String type, double money) {
 		String message = null; 
 
 		if (type.equals("Orange")) {
